requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-jose[cryptography]
python-multipart
python-dotenv
pytest
httpx
slowapi
